Fix tests marked RunnableOnService that aren't
Tests must make exclusively assertions about the contents of
PCollections to be marked RunnableOnService. Tests making assertions
about the client result cannot be ported between runners and do not meet
this requirement.

RunnableOnService tests that expect a specific exception as a result of
the pipeline failing are not portable. Remove the RunnableOnService
category from DoFnTest, DoFnWithContextTest, and WithTimestampTest tests
which expect exceptions.