Use input type in coder inference for MapElements and FlatMapElements
Previously, the input TypeDescriptor was unknown, so we would fail
to infer a coder for things like MapElements.of(SimpleFunction<T, T>)
even if the input PCollection provided a coder for T.

Now, the input type is plumbed appropriately and the coder is inferred.

This required internal changes to explicitly support good display data.
While doing this, I just added display data to SimpleFunction by analogy
with DoFn.