Prototype of Metrics API for Beam Java
This is includes a simple Counter metric and a Distribution that
supports SUM, COUNT, MIN, MAX and MEAN.

This is labeled @Experimental since it is only supported by the Direct
Runner and may change as it is implemented in other runners.

Future work includes adding additional kinds of metrics and wiring them
up to the various runners, as well as enriching the information
available from the Distribution metric.