Generate PubSub Subscriptions during Pipeline Execution
Instead of generating a random subscription at apply-time, create one
as required.

If splitIntoBundles is exercised, generate a subscription if there is
none specified, and create a new PubSub source that uses that
subscription. This is otherwise identical to the current behavior.

If splitIntoBundles is not exercised, and createReader is called with no
checkpoint, create a random subscription. Store a reference to this
subscription is the PubSub checkpoint, including when materialized.
This changes the wire format of PubSub checkpoints, and
as such this change is not update-compatible.