[BEAM-65] ProcessFn: create invoker in setup()
This makes more sense than creating it in constructor, since
the fn can be serialized (and will be - by Dataflow runner).

However, since direct runner creates the fn explicitly during
evaluation, we need to also call setup() explicitly.

Also, adds a getter for fn (a runner may need it to create a proper
ProcessContext when creating the SplittableProcessElementInvoker).