[BEAM-65] ProcessFn: support setup/teardown
Previously, ProcessFn did not explicitly invoke the underlying fn's
@Setup and @Teardown methods - it was assuming that those methods would
get invoked on that fn externally. This was true in direct runner,
but is not necessarily true in other runners: e.g., Dataflow runner
will serialize the whole ProcessFn and treat it mostly as a regular
DoFn, so it makes more sense to have lifecycle methods of ProcessFn
delegate to the underlying fn.

Also, adds a getter for fn (a runner may need it to create a proper
ProcessContext when creating the SplittableProcessElementInvoker).