Remove PValue.getProducingTransformInternal
This exposes too much information on the structure of the Graph to be
appropriate for non-graph related APIs. It also is not stable, and
subject to change at any point due to Pipeline Surgery APIs.

Update DataflowRunner to use the visited TransformHierarchy to obtain
the producing transforms for PValues, and store these values to
construct the graph.