Specialize ISM View Translations
These translations have a different coder than the original view, which
is required on the worker side to be transmitted as part of the
Singleton Output, and are generally not compatible with the replacements
performed during Pipeline Surgery. Specializing this translation allows
the updated coders to be set on the output view based on the overriding
transform, rather than the output node, as the view needs to remain
consistent.