Make Pubsub ackBatch safer
Checkpoints can be finalized at any point, including while the reader is
being closed or after it has been. When ackBatch is called and the
reader is closed, it should return. When close is called while a
different thread is using the client, it should wait for that thread to
complete before closing the reader.