Add SdkComponents
This takes SDK objects and assigns IDs to them. It is effectively a
ComponentsBuilder context where a component is referred to by the
Java object which is being translated, rather than by an opaque
string or protocol buffer.

The Builder for SdkComponents exposes mutation methods to add
components. The actual SdkComponents implementation is deeply immutable,
to signal that during actual conversion to/from protos should assume
that all elements are already recorded as a component.