Do not use infinite lateness in Reshuffle
Elements will be dropped only if the window they are assigned to is
expired, which can generally be performed arbitrarily.

Using infinite allowed lateness can cause an IllegalStateException if
the ReduceFnRunner observes certain combinations of timers and elements,
and attempts to buffer an element, setting a watermark hold for window
expiry (which is past the end of time).