Allow usage of scala 2.11 dependencies as well as cross compilation for flink runner
Flink allows being built against scala 2.11. But the Flink Runner did
not.

This commit alleviates that, as well as allowing for ensuring that
builds work against scala 2.11 dependencies. It introduces a
scala.version mvn property that is set to 2.10 as a default, as well as
a mvn profile that overrides the scala version to 2.11.