[BEAM-2623] Introduces Watch transform
The transform watches for new elements in a family of growing sets.
See design at http://s.apache.org/beam-watch-transform

As part of the implementation, I found and fixed a bug in tracking the
watermark in OutputAndTimeBoundedSplittableProcessElementInvoker.
The watermark must be captured at the moment checkpoint is taken,
because it describes timestamps of elements output from the checkpoint.

I also made direct runner by default checkpoint SDF's every 100 elements
rather than every 10000, to make it more aggressive - that's what
uncovered the bug above.