Integration test for Python HDFS implementation.
The test consists of running wordcount and writing the results to HDFS.
A Docker image is created containing a pseudo-distributed installation
of Hadoop and a Conda+pip virtual environment containing the Python Beam
SDK (built from the repo) and dependencies. The container command
starts Hadoop services and runs wordcount.

It will run as part of Python postcommit, since it takes about 30s-2m to
rebuild the image run when code changes.