Integration test for Python HDFS implementation.
The test consists of running wordcount and writing the results to HDFS.
A Docker image is created containing a pseudo-distributed installation
of Hadoop and a Conda+pip virtual environment containing the Python Beam
SDK (built from the repo) and dependencies. The container command
starts Hadoop services and runs wordcount.

It will run as part of Python postcommit, since it takes at least 67s to
run when code changes, with Docker caching the previous stages.