Add an ArtifactRetrievalService interface
Add a Local Artifact Retrieval Service which implements this interface.

The runner is not expected to interact directly with an instance of the
ArtifactRetrievalService, and is only exepcted to make it available to
environments it creates. As such, the interface is merely a signifier
to use within GrpcFnServer.

TODO: Implement