Introduce hooks for runner behavior.
A hook allows the runner to request the harness to call into
it to influence behavior of pipeline execution.

The currently identified hooks are

session recording
trace capturing
gRPC configuration

For a runner to use a hook, the registration method for the hook
must be called prior to beam.Init(). The request method for the hook
modifies the options for a pipeline prior to the runner's Execute()
method. If a registered hook is requested for a pipeline, the harness
will invoke the runner's supplied hook as needed based on the
specification of the hook.