Add a Primitive Impulse PTransform
This produces a single empty byte array, and is appropriate to use as
the root of a Pipeline instead of a Read transform. The composite
representation of a transform to produce elements at the root of a graph
is a chained Impulse -> ParDo.

Impulse is for SDK and Runner use, and should not be used in user
pipelines.

Currently restrict visibility of all constructors of Impulse, as there
are no current users. The presence of Impulse in the Core SDK is to
enable runners-core-construction to reference it appropriately during
translation time.