BEAM-3848: Adds retrying when SOLR writes are rejected.
This adds configuration to tune the behavior of writes to SOLR when failures occur.  The caller can provide max attempts and max duration for retrying.
Default behavior should be unchanged, and the code follows the same pattern as other IOs (e.g. JdbcIO) for the expontential backoff.

Please note that the WriteFn.flushBatch() signature was modified to allow unit testing.