BEAM-3894: Add Protocol Buffer support
This allows the Go SDK to use protocol buffer messages as a concrete
type. This creates a compile-time dependency on the protocol buffer
libraries. Internally, to build the SDK, you'd need protobuf support
to run "go generate" but this is a more direct dependency.

Adding a registration hook could move this dependency to be handled
differently, but this dep is lightweight and should be OK for the
immediate future.