[BEAM-3909] Add tests for Flink DoFnOperator side-input checkpointing
We test both checkpointing of pushed-back data and of side-input data.

This extends `PushbackSideInputDoFnRunner` to allow resetting any cached
information we keep about ready windows. Otherwise, it will not consider
side inputs that might have become ready in the same bundle.

Other than that, this only adds tests. The structure of the tests is:
 - Feed in some data
 - Checkpoint the operator and throw it away
 - Restore the operator
 - Send in some new data and verify that we get what we expect

In the first two tests we first send in side-input data, which tests
that side-inputs are correctly checkpointed. In the second set of tests
we first send in main-input data, which is retained because side inputs
are not ready.