[BEAM-3513] Removing PrimitiveCombineGroupedValues override w/ FnAPI.
When using the FnAPI, make it so that Combine.GroupedValues are not
replaced by the Dataflow specific PTransforms. The reason for this is
to allow the Combines to be translated to a Dataflow pipeline as the
constituent parts of the composite transform (in other words, send
Dataflow a GBK and a ParDo, instead of a CombineGroupedValues black
box).

This does not fully accomplish the final goal of portable combines since
this method doesn't do anything to help Combiner lifting, but it works
for the case of the basic Combine that has been implemented. It
successfully treats the Combine as a GBK and ParDo when not lifted.

Additional changes will need to be made to detect when we want to lift
Combines and in those situations handle that properly, which may mean
putting the Combines in the black box.