Convert some tests from ValidatesRunner -> NeedsRunner
This is an effort to reduce the overhead of our ValidatesRunner test
suite. By manual analysis of the @ValidatesRunner test, I've identified
a number of tests which don't seem to validate runner-specific
functionality and are better suited as @NeedsRunner tests. As a result,
these tests will only run on a single runner (DirectRunner), which
executes efficiently in-memory.