[BEAM-2915] Add support for handling bag user state to the java-fn-execution library to support runner integration.
Note that we specifically use an optimization which significantly simplifies the logical stream handling during get/append since we don't need
to encode/decode elements to actually find the element boundaries. We just store and replay the chunks the user gave us significantly
reducing the amount of pipeline manipulation the needs to happen.