[BEAM-4798] Fix UnboundedSourceWrapperTest
Before, the test did not try running the source with different simulated
subtask indices, i.e. we were not exercising many of the code paths.

This also fixes a bunch of concurrency issues in the source and test
that came up while extending it for simulating running at different
subtask indices.