[BEAM-4130] Build Docker image for Flink's JobServer
This adds a new Gradle module flink-job-server-container which builds a docker
image during the `docker` task. The image contains the FlinkJobServerDriver
which is the entry point for submitting Beam pipelines to the cluster.

The image can then be used to spawn a JobServer container when executing a Beam
pipeline, i.e. `p.run()`. The SDKs (Java/Python/Go) need to be updated to either
spawn up a container or use the address of a remote JobServer.