[BEAM-4130] Bring up Job Server container for Python jobs
This starts the Job Server in a container when Python pipelines are run and no
Job Server endpoint has been specified. The Job Server has the docker binaries
and socket mounted from the host to bring up SDK harness containers directly on
the host. This feature is currently only is available for Flink.

Due to host networking not being available on MacOs, we need to explicitly map
ports from the host system to the container. On Linux this is not necessary
because host networking works as expected. Generally, getting Docker support to
work on MacOs proved to be a bit tricky.

You can test this via:

```
 gradle :beam-sdks-python:portableWordCount
```

Some rough edges on MacOS:

- At times, the JobService socket is not available, even though
  `grpc.channel_ready_future(channel).result()` in the portable_runner returns.
  => Had to implement retry logic for the first request

- The port range for the SDK harness is restricted to be between 8100 and
  8200 (configurable) due to the necessary explicit port forwarding

Potentially problematic on Linux:

- The Docker binary might not be statically compiled and depend on shared
  libraries, though this needs to be tested