[BEAM-5250] Fix GlobalWindow's max timestamp in Python SDK
This fixes the Python streaming wordcount failing with an
`IllegalStateException` in the Runner code:

    "TimestampCombiner moved element from MAX_TIMESTAMP to earlier time END_OF_GLOBAL_WINDOW"

The Python SDK would create a `GlobalWindow` with a maximum timestamp of
MAX_TIMESTAMP which in turn would shift the 'element hold' in the Runner code
backwards. This is because the Runner's `GlobalWindow` maximum timestamp is
smaller than the Python SDK's.